First Verson
This is no longer working with Twitters API 1.1,

The code use to read a users twitter time line in an rss format, pull
out and edit some infromation for display.

Save it to a folder/file (cache/twitter_feed.html)

Check how old the file it, and if less than an hour old display that
file.

If more than an hour old re-do the file with the latest information and
then display it.